- Progress

+++ MotionCam-Android/MotionCam/src/main/java/com/motioncam/processor/ProcessorService.java
@@ -3,24 +3,32 @@ package com.motioncam.processor;
 import android.app.IntentService;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
+import android.content.ContentResolver;
+import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.BitmapFactory;
+import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
-import android.os.PowerManager;
+import android.os.ParcelFileDescriptor;
 import android.os.ResultReceiver;
+import android.provider.MediaStore;
 import android.util.Log;
 
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.core.app.NotificationCompat;
 
 import com.motioncam.R;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtil;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.concurrent.Callable;
 
@@ -35,9 +43,11 @@ public class ProcessorService extends IntentService {
     public static final String NOTIFICATION_CHANNEL_ID          = "MotionCamNotification";
 
     static class ProcessFile implements Callable<Void>, NativeProcessorProgressListener {
+        private final Context mContext;
         private final File mPendingFile;
         private final File mTempFileJpeg;
         private final File mTempFileDng;
+        private final File mOutputDirectory;
         private final File mOutputFileJpeg;
         private final File mOutputFileDng;
         private final NotificationManager mNotifyManager;
@@ -55,7 +65,8 @@ public class ProcessorService extends IntentService {
             return filename;
         }
 
-        ProcessFile(Context context, File pendingFile, File tempPath, File outputPath, boolean deleteAfterProcessing, ResultReceiver receiver) {
+        ProcessFile(Context context, File pendingFile, File tempPath, boolean deleteAfterProcessing, ResultReceiver receiver) {
+            mContext = context;
             mNativeProcessor = new NativeProcessor();
             mReceiver = receiver;
             mPendingFile = pendingFile;
@@ -68,8 +79,12 @@ public class ProcessorService extends IntentService {
             mTempFileJpeg = new File(tempPath, outputFileNameJpeg);
             mTempFileDng = new File(tempPath, outputFileNameDng);
 
-            mOutputFileJpeg = new File(outputPath, outputFileNameJpeg);
-            mOutputFileDng = new File(outputPath, outputFileNameDng);
+            // Set up the output path to point to the camera DCIM folder
+            File dcimDirectory = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM);
+
+            mOutputDirectory = new File(dcimDirectory, "Camera");
+            mOutputFileJpeg = new File(mOutputDirectory, outputFileNameJpeg);
+            mOutputFileDng = new File(mOutputDirectory, outputFileNameDng);
 
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 NotificationChannel notificationChannel = new NotificationChannel(
@@ -114,13 +129,66 @@ public class ProcessorService extends IntentService {
             else
                 mPendingFile.renameTo(new File(mPendingFile.getPath() + ".complete"));
 
-            // Copy to output folder
-            if(mTempFileDng.exists())
-                FileUtils.copyFile(mTempFileDng, mOutputFileDng);
+            // Copy to media store
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+                if (mTempFileDng.exists())
+                    saveToMediaStore(mTempFileDng);
 
-            FileUtils.copyFile(mTempFileJpeg, mOutputFileJpeg);
+                if (mTempFileJpeg.exists())
+                    saveToMediaStore(mTempFileJpeg);
+            }
+            // Legacy copy file
+            else {
+                if(!mOutputDirectory.exists()) {
+                    if(!mOutputDirectory.mkdirs()) {
+                        Log.e(TAG, "Failed to create " + mOutputDirectory);
+                        throw new IOException("Failed to create output directory");
+                    }
+                }
+
+                if(mTempFileDng.exists())
+                    FileUtils.copyFile(mTempFileDng, mOutputFileDng);
+
+                if(mTempFileJpeg.exists())
+                    FileUtils.copyFile(mTempFileJpeg, mOutputFileJpeg);
+            }
 
             return null;
+        }
+
+        @RequiresApi(api = Build.VERSION_CODES.Q)
+        private void saveToMediaStore(File inputFile) throws IOException {
+            ContentResolver resolver = mContext.getApplicationContext().getContentResolver();
+
+            Uri imageCollection;
+
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+                imageCollection = MediaStore.Images.Media.getContentUri(MediaStore.VOLUME_EXTERNAL_PRIMARY);
+            } else {
+                imageCollection = MediaStore.Images.Media.EXTERNAL_CONTENT_URI;
+            }
+
+            ContentValues imageDetails = new ContentValues();
+
+            imageDetails.put(MediaStore.Images.Media.DISPLAY_NAME,  inputFile.getName());
+            imageDetails.put(MediaStore.Images.Media.MIME_TYPE,     "image/jpeg");
+            imageDetails.put(MediaStore.Images.Media.DATE_ADDED,    System.currentTimeMillis());
+            imageDetails.put(MediaStore.Images.Media.DATE_TAKEN,    System.currentTimeMillis());
+            imageDetails.put(MediaStore.Images.Media.RELATIVE_PATH, Environment.DIRECTORY_DCIM);
+            imageDetails.put(MediaStore.Images.Media.IS_PENDING,    1);
+
+            Uri imageContentUri = resolver.insert(imageCollection, imageDetails);
+
+            try (ParcelFileDescriptor pfd = resolver.openFileDescriptor(imageContentUri, "w", null)) {
+                FileOutputStream outStream = new FileOutputStream(pfd.getFileDescriptor());
+
+                IOUtil.copy(new FileInputStream(inputFile), outStream);
+            }
+
+            imageDetails.clear();
+            imageDetails.put(MediaStore.Images.Media.IS_PENDING, 0);
+
+            resolver.update(imageContentUri, imageDetails, null, null);
         }
 
         @Override
@@ -188,17 +256,7 @@ public class ProcessorService extends IntentService {
 
         Log.i(TAG, "Found " + pendingFiles.length + " images to process");
 
-        File dcimDirectory = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM);
-        File outputDirectory = new File(dcimDirectory, "Camera");
         File tmpDirectory = new File(getFilesDir(), "tmp");
-
-        // Create final output directory
-        if(!outputDirectory.exists()) {
-            if(!outputDirectory.mkdirs()) {
-                Log.e(TAG, "Failed to create " + outputDirectory);
-                return;
-            }
-        }
 
         // Create temporary directory
         if(!tmpDirectory.exists()) {
@@ -209,26 +267,15 @@ public class ProcessorService extends IntentService {
         }
 
         // Process all files
-        PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
-        PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "MotionCam::ProcessingLockTag");
+        for(File file : pendingFiles) {
+            ProcessFile processFile = new ProcessFile(getApplicationContext(), file, tmpDirectory, deleteAfterProcessing, receiver);
 
-        try
-        {
-            wakeLock.acquire();
-
-            for(File file : pendingFiles) {
-                ProcessFile processFile = new ProcessFile(getApplicationContext(), file, tmpDirectory, outputDirectory, deleteAfterProcessing, receiver);
-
-                try {
-                    processFile.call();
-                }
-                catch (IOException e) {
-                    e.printStackTrace();
-                }
+            try {
+                processFile.call();
             }
-        }
-        finally {
-            wakeLock.release();
+            catch (IOException e) {
+                e.printStackTrace();
+            }
         }
     }
 }
